Conflicts:
	lib/nokogiri/css/tokenizer.rex
	lib/nokogiri/xml/document.rb
	lib/nokogiri/xml/node.rb
	test/helper.rb
	test/html/test_element_description.rb
	test/test_xslt_transforms.rb
	test/xml/test_node.rb

+++ ext/nokogiri/xml_io.c
@@ -10,7 +10,7 @@ int io_read_callback(void * ctx, char * buffer, int len) {
 
   memcpy(buffer, StringValuePtr(string), (unsigned int)RSTRING_LEN(string));
 
-  return RSTRING_LEN(string);
+  return (int)RSTRING_LEN(string);
 }
 
 int io_write_callback(void * ctx, char * buffer, int len) {
